chore: move formsLib module
