+++ src/components/form-contents/configured-form-contents.jsx
@@ -2,7 +2,7 @@ import { reduxForm } from 'redux-form';
 
 import FormContentsComponent from './form-contents.component';
 import validate from './form-contents-validations';
-import { asyncValidateDatasetInvokePatch } from '../../entrypoints/main/router/datasets/pages/dataset-page/formsLib/asyncValidateDatasetInvokePatch';
+import { asyncValidateDatasetInvokePatch } from '../dataset-registration-form/formsLib/asyncValidateDatasetInvokePatch';
 
 const config = {
   form: 'contents',
