+++ src/components/form-distribution/form-distribution.js
@@ -5,7 +5,7 @@ import _throttle from 'lodash/throttle';
 import FormDistributionPure from './form-distribution-pure';
 import { distributionTypes } from './distribution-types';
 import validate from './form-distribution-validations';
-import { asyncValidateDatasetInvokePatch } from '../../entrypoints/main/router/datasets/pages/dataset-page/formsLib/asyncValidateDatasetInvokePatch';
+import { asyncValidateDatasetInvokePatch } from '../dataset-registration-form/formsLib/asyncValidateDatasetInvokePatch';
 
 const setInitialValues = withProps(({ datasetItem: { distribution } }) => ({
   initialValues: {
