+++ src/components/form-informationmodel/external-information-models/external-information-models.component.tsx
@@ -8,7 +8,7 @@ import InputField from '../../fields/field-input/field-input.component';
 import MultilingualField from '../../multilingual-field/multilingual-field.component';
 import LinkReadonlyField from '../../fields/field-link-readonly/field-link-readonly.component';
 import InputFieldReadonly from '../../fields/field-input-readonly/field-input-readonly.component';
-import { datasetFormPatchThunk } from '../../../entrypoints/main/router/datasets/pages/dataset-page/formsLib/asyncValidateDatasetInvokePatch';
+import { datasetFormPatchThunk } from '../../dataset-registration-form/formsLib/asyncValidateDatasetInvokePatch';
 
 import '../form-informationmodel.component.scss';
