+++ src/components/form-publish/form-publish-configurer.js
@@ -1,5 +1,5 @@
 import { reduxForm } from 'redux-form';
-import { asyncValidateDatasetInvokePatch } from '../../entrypoints/main/router/datasets/pages/dataset-page/formsLib/asyncValidateDatasetInvokePatch';
+import { asyncValidateDatasetInvokePatch } from '../dataset-registration-form/formsLib/asyncValidateDatasetInvokePatch';
 
 export const formPublishConfigurer = reduxForm({
   form: 'datasetPublish',
