+++ src/components/form-reference/configured-form-reference.jsx
@@ -2,7 +2,7 @@ import { reduxForm } from 'redux-form';
 
 import FormReference from './form-reference.component';
 import validate from './form-reference-validations';
-import { asyncValidateDatasetInvokePatch } from '../../entrypoints/main/router/datasets/pages/dataset-page/formsLib/asyncValidateDatasetInvokePatch';
+import { asyncValidateDatasetInvokePatch } from '../dataset-registration-form/formsLib/asyncValidateDatasetInvokePatch';
 
 const config = {
   form: 'reference',
