+++ src/components/form-spatial/configured-form-spatial.jsx
@@ -2,7 +2,7 @@ import { reduxForm } from 'redux-form';
 
 import FormSpatial from './form-spatial.component';
 import validate from './form-spatial-validations';
-import { asyncValidateDatasetInvokePatch } from '../../entrypoints/main/router/datasets/pages/dataset-page/formsLib/asyncValidateDatasetInvokePatch';
+import { asyncValidateDatasetInvokePatch } from '../dataset-registration-form/formsLib/asyncValidateDatasetInvokePatch';
 
 const config = {
   form: 'spatial',
