+++ src/components/form-type/configured-form-type.jsx
@@ -1,7 +1,7 @@
 import { reduxForm } from 'redux-form';
 
 import FormType from './form-type.component';
-import { asyncValidateDatasetInvokePatch } from '../../entrypoints/main/router/datasets/pages/dataset-page/formsLib/asyncValidateDatasetInvokePatch';
+import { asyncValidateDatasetInvokePatch } from '../dataset-registration-form/formsLib/asyncValidateDatasetInvokePatch';
 
 const config = {
   form: 'type',
